refactor onsubmit code
